Added version cast
Provides an elegant way in which to add the `Version` type to `Kernel`
so `Version` is available much like other casting functions are.
Examples: `Integer`, `String`, `Array`, `Hash`, etc.

Usage:

    using Versionaire::Cast

    Version "1.2.3"
    # => #<struct Versionaire::Version major=1, minor=2, patch=3>

The Reek violations are being ignored because Reek doesn't handle
conversion functions or refinements all that well at the moment. I've
opened a few related issues with Reek to pursue this further:

- [Utility Function - Kernel](https://chl.li/crVLS)
- [Utility Function - Class](https://chl.li/oefGg)