Added version inspection
Necessary to provide an *escaped* string representation of the version
when inspecting it. This makes the version more compatible with
existing inspections of primitives such as strings, arrays, hashes,
floats, regular expressions, etc.

The `#to_a` and `#to_h` specs were refactored as part of this work in
an effort to slightly improve the alpha-sorting of them.